@@ -34,9 +34,10 @@ module LiquidFilter
     obj.image(size)
   end
 
-  def variant_options_values(variant)
-      names = variant.options.collect{|option| ["#{option.option_type.name}: #{option.value}"]}
-      names.join(' - ')
+  def variant_options_values(option)
+      # names = options.collect{|option| ["#{option.option_type.name}: #{option.value}"]}
+      # names.join(' - ')
+      "#{option.option_type.name}: #{option.value}"
   end
 
   def form_token(token)
