@@ -74,8 +74,8 @@ module Imentore
           flash[:alert] = @order.errors.full_messages.join
           redirect_to cart_path
         elsif @order.chargeable? and @order.save
-          @order.place
           charge
+          @order.place
         else
           flash[:alert] = @order.errors.full_messages.join(" / ")
           render :confirm
@@ -132,6 +132,7 @@ module Imentore
     end
 
     def sync_pg
+      binding.pry
       current_store = Imentore::Store.find(params[:store_id])
       notification_code = {notificationCode: params[:notificationCode]}
       provider_class = current_store.payment_methods.find_by_handle('pag_seguro').provider
