@@ -25,7 +25,7 @@ module Imentore
 
 
                    },
-        "total_items" =>  @cart.items.size,
+        "total_items" =>  @cart.total_items,
         "total_amount" => number_with_price(@cart.amount),
         "message" => {
                        "success" => I18n.t(:add, scope: 'helpers.cart.create')
